Reworked serverSetup.js
The biggest change is to the way arguments are handled, now arguments are treated like a regular array (Which they are, the script has complete access to modify and read it). This means that the flags no longer need to be in a certain order.
I also changed the input for script copying to be a prompt rather than an argument, since there's no real cost to doing so.

I also changed some other aspects of the code such as replacing the port buster variables with has(PortName) rather than the name of the port buster exe.

Changed the info color from the default color to white, side note; I encourage you to change these colors to suite your color theme for BitBurner if you use this script as these colors could be hard to see for you.

Added the --ls flag to clean up the output while still allowing the player to see what servers have been hacked.